requirements.txt
google-cloud-aiplatform
langchain
chromadb
pytube
youtube-transcript-api
Flask
langchain-google-vertexai